backend/requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
geoalchemy2==0.14.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.6.0
python-dotenv==1.0.0
redis==5.0.1
requests==2.31.0
supabase==1.2.0 